more tests, cleanup
